+++ spec/internal/app/controllers/sessions_controller.rb
@@ -1,0 +1,11 @@
+class SessionsController < ApplicationController
+  def create
+    session[:status] = params[:status]
+    head :status => 200
+  end
+
+  def destroy
+    reset_session
+    head :status => 200
+  end
+end