+++ spec/internal/config/initializers/session.rb
@@ -1,1 +1,0 @@
-Combustion::Application.config.session_store SequelRails::SessionStore