+++ spec/internal/config/initializers/session_store.rb
@@ -1,0 +1,1 @@
+Combustion::Application.config.session_store :sequel_store